requirements.txt
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
xgboost>=1.7
joblib>=1.2

torch>=2.0
torchvision>=0.15
pillow>=9.5

matplotlib>=3.7
